src/requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
model-bakery==1.4.0
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.0
pytest-django==4.5.2
python-dotenv==0.19.2
python-magic==0.4.25
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.7.0
